Cleaning up some code in the importer and metaprogramming some logic
While I understand that there is a degree of context that is lost in the
metaprogramming solution that I've provided, I enjoy metaprogramming for
its artistry and would like to show off some of my abilities in this
regard. Normally I wouldn't do this in code that is interacted with more
readily, but I enjoy doing this for fun. If you're not having fun while
programming, in my opinion, you're doing something wrong.